Improvements to CloneableObject
Made CloneableObject something more customizable with transformation
methods that can be subclassed and creation methods that can be
sublcassed, and made DictAdaptor that subclasses from CloneableObject
and made CloneableWithCopyOptions that uses the transform to allow for
copy transforms and such. Added comments for how to subclass from both
CloneableWithCopyOptions and DictAdaptor, but have not made that yet.